mini/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Limiter
psycopg2-binary
python-dotenv